Sort function is improved: size of array is amended from int to size_t

+++ bubblesort.h
@@ -1,2 +1,2 @@
 #pragma once
-void bubbleSort(int* ar, int n);
+void bubbleSort(int* ar, size_t n);
